Fix PolymorphicPtr destruction bug and finalize list logic and commands
Resolved a critical issue in the PolymorphicPtr destructor logic that caused premature deletion of dynamically allocated objects after use.

The problem stemmed from incorrect ownership semantics in finder functions. To fix it, the return types of finder-related functions were adjusted to prevent unsafe deletions.

Minor logic adjustments were made to support safe memory handling.

Finalised implementation of the remaining list-related logic and associated command execution.

Performed general bug fixes across the system to ensure correct flow and stability.

+++ Core/header/Ticket.h
@@ -5,18 +5,18 @@ class Ticket
 {
 public:
 	Ticket();
-	Ticket(int movieId, int row, int col);
+	Ticket(int movieId, size_t row, size_t col);
 
 	int getMovieId() const;
-	int getRow() const;
-	int getCol() const;
+	size_t getRow() const;
+	size_t getCol() const;
 
 	void saveToBinaryFile(std::ofstream& ofs) const;
 	void loadFromBinaryFile(std::ifstream& ifs);
 
 private:
 	int movieId;
-	int row;
-	int col;
+	size_t row;
+	size_t col;
 };
